requirements.txt
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorlog==4.0.2
Flask==1.0.3
Flask-HTTPAuth==3.3.0
geographiclib==1.49
geopy==1.20.0
gevent==1.4.0
google-api-core==1.13.0
google-auth==1.6.3
google-cloud-core==1.0.2
google-cloud-pubsub==0.42.1
google-cloud-storage==1.16.1
google-cloud-videointelligence==1.9.0
google-resumable-media==0.3.2
googleapis-common-protos==1.6.0
greenlet==0.4.15
grpc-google-iam-v1==0.11.4
grpcio==1.22.0
idna==2.8
inotify==0.2.10
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
nose==1.3.7
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pytz==2019.1
requests==2.22.0
rsa==4.0
sightengine==1.4.0
six==1.12.0
SQLAlchemy==1.4.18
urllib3==1.25.2
websocket-client==0.56.0
websockets==8.0
Werkzeug==0.15.4
flask-login
flask-cors
python-dateutil
gunicorn==20.0.4